% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateVariables.R
\name{aggregateVariables}
\alias{aggregateVariables}
\title{Report variables in relation to the vehicle fleet.}
\usage{
aggregateVariables(vars, mapAggregation, weight = NULL)
}
\arguments{
\item{helpers}{}
}
\description{
Variables like energy intensity and capital costs are linked to the
construction year of a vehicle.
As energy intensity and capital costs change over time for new sales, the composition
of the fleet from vehicles of different construction years needs to be taken into account
to report these variables.
}
\author{
Johanna Hoppe
}
