% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConvergence.R
\name{checkConvergence}
\alias{checkConvergence}
\title{Compare parameters in edget and REMIND and store their deviation to track model convergence}
\usage{
checkConvergence()
}
\description{
Compare parameters in edget and REMIND and store their deviation to track model convergence
}
\author{
Johanna Hoppe
}
