% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{checkForNAsAndDups}
\alias{checkForNAsAndDups}
\title{Check a data.table for NAs and duplicates and throw an error if needed}
\usage{
checkForNAsAndDups(dt, varname, codePosition)
}
\arguments{
\item{dt}{data.table to be checked}

\item{varname}{name of the variable}

\item{codePosition}{position in the code to find the bug}
}
\description{
Check a data.table for NAs and duplicates and throw an error if needed
}
\author{
Johanna Hoppe
}
