% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeOutput.R
\name{harmonizeOutput}
\alias{harmonizeOutput}
\title{Harmonize data to last REMIND iteration in coupled mode}
\usage{
harmonizeOutput(edgetOutputDir, baseVarSet, data)
}
\arguments{
\item{edgetOutputDir}{edgeTransport output directory}

\item{baseVarSet}{calculated base variable set including energy intensity and final energy on fleet level}

\item{data}{list of raw model output data}
}
\value{
The function returns a list with the harmonized energy intensity and final energy data
}
\description{
The two models are coupled via energy service demand, energy intensity and capital expenditure on CES node level.
As REMIND converges the output between EDGE-T and REMIND still slightly differs.
This function calculates harmonization factors that are applied on the EDGE-T energy intensity and final energy values.
Furthermore, the harmonization factors indicate how well the two models converged and can be used for analysis purposes.
}
\author{
Johanna Hoppe
}
