% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{prepareForREMIND}
\alias{prepareForREMIND}
\title{Prepare data to report to REMIND}
\usage{
prepareForREMIND(dt, demScen, SSPscen, transportPolScen)
}
\arguments{
\item{dt}{data.table with data to report to REMIND}

\item{demScen}{demand scenario}

\item{SSPscen}{SSP scenario}

\item{transportPolScen}{transport policy scenario}
}
\value{
data table ready to report to REMIND
}
\description{
Prepare data to report to REMIND
}
\author{
Johanna Hoppe
}
