% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{renameDuplicateVariables}
\alias{renameDuplicateVariables}
\title{Renames vertain variables to prevent duplicates for variables that are reported also in remind2}
\usage{
renameDuplicateVariables(vars)
}
\arguments{
\item{vars}{MIF variables}
}
\value{
MIF variables that are renamed to "Transport edge" where necessary
}
\description{
Renames vertain variables to prevent duplicates for variables that are reported also in remind2
}
\author{
Johanna Hoppe
}
