% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAnalyticsVarSet.R
\name{reportAnalyticsVarSet}
\alias{reportAnalyticsVarSet}
\title{Report analytics variable set}
\usage{
reportAnalyticsVarSet(data, timeResReporting)
}
\arguments{
\item{data}{List that contains the model results to report the analytics variable set}

\item{timeResReporting}{Timesteps to be reported}
}
\value{
Analytics variable set
}
\description{
Report analytics variable set
}
\author{
Johanna Hoppe
}
