% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportExtendedTransportVarSet.R
\name{reportExtendedTransportVarSet}
\alias{reportExtendedTransportVarSet}
\title{Report extended detailed transport variable set}
\usage{
reportExtendedTransportVarSet(data, baseVarSet, timeResReporting)
}
\arguments{
\item{data}{List that contains the model results to report the extended detailed transport variable set}

\item{baseVarSet}{List that contains the basic variables}

\item{timeResReporting}{Time resolution for variable reporting}
}
\value{
Extended detailed transport output variable set
}
\description{
Report extended detailed transport variable set
}
\author{
Johanna Hoppe
}
