% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportExtendedTransportVarSet.R
\name{reportExtendedVarSet}
\alias{reportExtendedVarSet}
\title{Report extended detailed transport variable set}
\usage{
reportExtendedVarSet(data, timeResReporting)
}
\arguments{
\item{data}{List that contains the model results to report the extended detailed transport variable set}

\item{vehiclesConstrYears}{}
}
\value{
Extended detailed transport output variable set
}
\description{
Report extended detailed transport variable set
}
\author{
Johanna Hoppe
}
