% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLiquidsAndGasesComposition.R
\name{reportLiquidsAndGasesComposition}
\alias{reportLiquidsAndGasesComposition}
\title{Report the split of liquids and gases into fossil|bio|hydrogen}
\usage{
reportLiquidsAndGasesComposition(dtFE, gdxPath, timeResReporting, helpers)
}
\arguments{
\item{dtFE}{Final energy data for liquids and gases}

\item{gdxPath}{Path to REMIND gdx, which contains the share of the various production routes
for liquid and gaseous energy carriers}

\item{timeResReporting}{Time resolution for variable reporting}

\item{helpers}{List of helpers}
}
\value{
Final energy for liquids and gases split into fossil|bio|hydrogen
}
\description{
Report the split of liquids and gases into fossil|bio|hydrogen
}
\author{
Johanna Hoppe
}
