% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportToREMINDcapitalCosts.R
\name{reportToREMINDcapitalCosts}
\alias{reportToREMINDcapitalCosts}
\title{Report to REMIND p35_esCapCost}
\usage{
reportToREMINDcapitalCosts(
  fleetCapCosts,
  fleetESdemand,
  hybridElecShare,
  timeResReporting,
  demScen,
  SSPscen,
  transportPolScen,
  helpers
)
}
\arguments{
\item{fleetCapCosts}{capital costs on fleet level}

\item{fleetESdemand}{energy service demand on fleet level}

\item{hybridElecShare}{share of electric driving for hybrid electric vehicles}

\item{timeResReporting}{time resolution reporting}

\item{demScen}{demand scenario}

\item{SSPscen}{SSP scenario}

\item{transportPolScen}{transport policy scenario}

\item{helpers}{list of helpers}
}
\value{
Capital Costs per CES node in [2005US$/pkm or 2005US$/tkm]
}
\description{
Report to REMIND p35_esCapCost
}
\author{
Johanna Hoppe
}
