% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportToREMINDesDemand.R
\name{reportToREMINDesDemand}
\alias{reportToREMINDesDemand}
\title{Report energy service demand to REMIND}
\usage{
reportToREMINDesDemand(
  fleetESdemand,
  hybridElecShare,
  timeResReporting,
  demScen,
  SSPscen,
  transportPolScen,
  helpers
)
}
\arguments{
\item{fleetESdemand}{energy service demand on fleet level}

\item{hybridElecShare}{share of electric driving for hybrid electric vehicles}

\item{timeResReporting}{time resolution reporting}

\item{demScen}{demand scenario}

\item{SSPscen}{SSP scenario}

\item{transportPolScen}{transport policy scenario}

\item{helpers}{list of helpers}
}
\value{
Energy service demand per technology in [trillion pkm/trillion tkm]
}
\description{
Note that this is a helpers function that accounts for structural differences in REMIND and EDGE-T.
The REMIND CES tree cannot represent active modes, as they deliver valuable energy service demand without capital cost and final energy production.
Just filtering out the active modes energy service demand allocated by EDGE-Transport would lead to inconsistencies between both models:
E.g. REMIND would split the whole trn_pass energy service demand according to the FE shares and energy efficiencies reported by EDGE-T (without considering the active modes)
and would calculate a different FE value compared to EDGE-T that reserves a part of the trn_pass energy service demand
to the active modes.
}
\details{
Solution: split and evenly distribute the energy service demand of active modes across all technologies to calculate energy efficiency and capital cost per pkm.
Consequence: This leads to the fact that in REMIND energy service demand, energy efficiencies, an capital costs on technology level are rather abstract values that are not technically correct.
Hence, they should not be reported on technology level by REMIND (only by EDGE-T). The FE values on technology level are represented correctly as well as the energy service demand on CES node level.
The decision of the technology share is anyway made by EDGE-T and not by REMIND.
}
\author{
Johanna Hoppe
}
