% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportToREMINDfinalEnergyDemand.R
\name{reportToREMINDfinalEnergyDemand}
\alias{reportToREMINDfinalEnergyDemand}
\title{Report to REMIND f35_demByTech}
\usage{
reportToREMINDfinalEnergyDemand(
  fleetFEdemand,
  timeResReporting,
  demScen,
  SSPscen,
  transportPolScen,
  helpers
)
}
\arguments{
\item{fleetFEdemand}{final energy demand on fleet level}

\item{timeResReporting}{time resolution reporting}

\item{demScen}{demand scenario}

\item{SSPscen}{SSP scenario}

\item{transportPolScen}{transport policy scenario}

\item{helpers}{list of helpers}
}
\value{
Final energy demand on CES level per transport fuel technology [TWa]
}
\description{
Report to REMIND f35_demByTech
}
\author{
Johanna Hoppe
}
