% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportToREMINDfinalEnergyShares.R
\name{reportToREMINDfinalEnergyShares}
\alias{reportToREMINDfinalEnergyShares}
\title{Report to REMIND final energy shares f35_shFeCes}
\usage{
reportToREMINDfinalEnergyShares(
  fleetFEdemand,
  timeResReporting,
  demScen,
  SSPscen,
  transportPolScen,
  helpers
)
}
\arguments{
\item{fleetFEdemand}{Final energy demand on fleet level}

\item{timeResReporting}{time resolution reporting}

\item{demScen}{demand scenario}

\item{SSPscen}{SSP scenario}

\item{transportPolScen}{transport policy scenario}

\item{helpers}{list of helpers}
}
\value{
Final energy shares of transport fuel technologies in [-]
}
\description{
Report to REMIND final energy shares f35_shFeCes
}
