% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportToREMINDtrpdemand.R
\name{reportToREMINDtrpdemand}
\alias{reportToREMINDtrpdemand}
\title{Report to REMIND p29_trpdemand}
\usage{
reportToREMINDtrpdemand(
  fleetESdemand,
  hybridElecShare,
  timeResReporting,
  demScen,
  SSPscen,
  transportPolScen,
  helpers
)
}
\arguments{
\item{fleetESdemand}{energy service demand on fleet level}

\item{hybridElecShare}{share of electric driving for hybrid electric vehicles}

\item{timeResReporting}{time resolution reporting}

\item{demScen}{demand scenario}

\item{SSPscen}{SSP scenario}

\item{transportPolScen}{transport policy scenario}

\item{helpers}{list of helpers}
}
\value{
Energy service demand per CES node in [trillion pkm/trillion tkm]
}
\description{
Report to REMIND p29_trpdemand
}
\author{
Johanna Hoppe
}
