% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTransportVarSet.R
\name{reportTransportVarSet}
\alias{reportTransportVarSet}
\title{Report detailed transport variable set}
\usage{
reportTransportVarSet(data, baseVarSet, timeResReporting)
}
\arguments{
\item{data}{List that contains the model results to report the detailed transport variable set}

\item{baseVarSet}{Basic output variable set}

\item{timeResReporting}{Timesteps to be reported}
}
\value{
Detailed transport output variable set
}
\description{
Report detailed transport variable set
}
\author{
Johanna Hoppe
}
