% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportUE.R
\name{reportUE}
\alias{reportUE}
\title{Report variables in relation to the vehicle fleet.}
\usage{
reportUE(FEdemand, helpers)
}
\arguments{
\item{FEdemand}{Finale energy demand}

\item{helpers}{List with helpers}
}
\value{
Useful energy demand
}
\description{
Report variables in relation to the vehicle fleet.
}
\author{
Johanna Hoppe
}
