% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storeData.R
\name{storeData}
\alias{storeData}
\title{Store EDGE-Transport model results}
\usage{
storeData(outputFolder, outputRaw = NULL, ...)
}
\arguments{
\item{outputFolder}{Path to folder for storing output data}

\item{outputRaw}{Raw model results}

\item{...}{Optional passing of additional variables}
}
\description{
This function creates the EDGE-Transport outputfolder and stores all outputfiles in the respective subfolders
}
\author{
Johanna Hoppe
}
