% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReportAggregatedCosts.R
\name{toolReportAggregatedCosts}
\alias{toolReportAggregatedCosts}
\title{Report cost variables in aggregated levels:
Capital costs sales, operating costs (total non-fuel), fuel costs}
\usage{
toolReportAggregatedCosts(combinedCAPEXandOPEX)
}
\arguments{
\item{combinedCAPEXandOPEX}{detailed data on cost variables}
}
\value{
aggregated cost variables
}
\description{
Report cost variables in aggregated levels:
Capital costs sales, operating costs (total non-fuel), fuel costs
}
\author{
Johanna Hoppe
}
