% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReportBaseVarSet.R
\name{toolReportBaseVarSet}
\alias{toolReportBaseVarSet}
\title{Report basic variable set needed to report REMIND input data and detailed transport data}
\usage{
toolReportBaseVarSet(data, timeResReporting)
}
\arguments{
\item{data}{List that contains at least the model results to report the basic variable set}

\item{timeResReporting}{Timesteps to be reported}
}
\value{
list of intensive and extensive output variables
}
\description{
Report basic variable set needed to report REMIND input data and detailed transport data
}
\author{
Johanna Hoppe
}
