% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReportEmissions.R
\name{toolReportEmissions}
\alias{toolReportEmissions}
\title{Report emissions allocated to the transport sector.
Only direkt emissions from liquid or gaseous energy carriers used in ICEs are considered.
Indirect emissions from electricity and hydrogen are not allocated to the transport sector.}
\usage{
toolReportEmissions(dtFE, gdxPath, prefix, helpers)
}
\arguments{
\item{dtFE}{Final energy data for liquids and gases}

\item{gdxPath}{Path to REMIND fulldata.gdx containing emission factors}

\item{prefix}{Prefix that specifies the emissions we are referring to in the variable name
(either tailpipe or demand)}

\item{helpers}{}
}
\value{
Emissions data for provided values in dtFE
}
\description{
Report emissions allocated to the transport sector.
Only direkt emissions from liquid or gaseous energy carriers used in ICEs are considered.
Indirect emissions from electricity and hydrogen are not allocated to the transport sector.
}
\author{
Johanna Hoppe
}
