% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReportFE.R
\name{toolReportFE}
\alias{toolReportFE}
\title{Report variables in relation to the vehicle fleet.}
\usage{
toolReportFE(
  fleetEnergyIntensity,
  fleetESdemand,
  loadFactor,
  hybridElecShare,
  helpers
)
}
\arguments{
\item{fleetEnergyIntensity}{Energy intensity linked to the vehicle fleet}

\item{fleetESdemand}{Energy service on technology level linked to the vehicle fleet}

\item{loadFactor}{Persons or tons per vehicle}

\item{hybridElecShare}{Share of electric driving in hybrid electric vehicles}

\item{helpers}{list of helpers}
}
\value{
Final energy consumed by all modes and technologies
}
\description{
Variables like energy intensity and capital costs are linked to the
construction year of a vehicle.
As energy intensity and capital costs change over time for new sales, the composition
of the fleet from vehicles of different construction years needs to be taken into account
to report these variables.
}
\author{
Johanna Hoppe
}
