% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReportFleetVariables.R
\name{toolReportFleetVariables}
\alias{toolReportFleetVariables}
\title{Report variables in relation to the vehicle fleet.}
\usage{
toolReportFleetVariables(salesData, vehiclesConstrYears, helpers)
}
\arguments{
\item{salesData}{variables linked to new sales in the respective years}

\item{vehiclesConstrYears}{vehicle numbers and their construction years for the respective years}

\item{helpers}{list of helpers}
}
\description{
Variables like energy intensity and capital costs are linked to the
construction year of a vehicle.
As energy intensity and capital costs change over time for new sales, the composition
of the fleet from vehicles of different construction years needs to be taken into account
to report these variables.
}
\author{
Johanna Hoppe
}
