% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReportMIF.R
\name{toolReportMIF}
\alias{toolReportMIF}
\title{Report variables in relation to the vehicle fleet.}
\usage{
toolReportMIF(
  vars,
  GDPMER,
  helpers,
  scenario,
  model,
  gdx,
  reportExtendedTransportData = FALSE
)
}
\arguments{
\item{vars}{Variables to be aggregated and converted into MIF format}

\item{GDPMER}{GDP on market exchange rate basis to be used as weight for regional aggregation}

\item{helpers}{List of helpers}

\item{scenario}{Scenario name in MIF entry}

\item{model}{Model name in MIF entry}

\item{gdx}{GDX file containing further regional aggregation levels}

\item{reportExtendedTransportData}{Switch to enable the extended transport variable set}
}
\value{
Variables provided in different aggregation levels in MIF format
}
\description{
Variables like energy intensity and capital costs are linked to the
construction year of a vehicle.
As energy intensity and capital costs change over time for new sales, the composition
of the fleet from vehicles of different construction years needs to be taken into account
to report these variables.
}
\author{
Johanna Hoppe
}
