% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReportREMINDinputDataVarSet.R
\name{toolReportREMINDinputDataVarSet}
\alias{toolReportREMINDinputDataVarSet}
\title{Report detailed transport variable set}
\usage{
toolReportREMINDinputDataVarSet(
  fleetESdemand,
  fleetFEdemand,
  fleetEnergyIntensity,
  fleetCapCosts,
  combinedCAPEXandOPEX,
  scenSpecPrefTrends,
  scenSpecEnIntensity,
  initialIncoCosts,
  annualMileage,
  timeValueCosts,
  demScen,
  SSPscen,
  transportPolScen,
  helpers
)
}
\arguments{
\item{fleetESdemand}{xx}

\item{fleetFEdemand}{xx}

\item{fleetEnergyIntensity}{xx}

\item{fleetCapCosts}{xx}

\item{combinedCAPEXandOPEX}{xx}

\item{scenSpecPrefTrends}{xx}

\item{scenSpecEnIntensity}{xx}

\item{initialIncoCosts}{description}

\item{annualMileage}{description}

\item{timeValueCosts}{description}

\item{demScen}{description}

\item{SSPscen}{description}

\item{transportPolScen}{description}

\item{helpers}{description}
}
\value{
Detailed transport output variable set
}
\description{
Report detailed transport variable set
}
\author{
Johanna Hoppe
}
