% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReportREMINDinputDataVarSet.R,
%   R/toolReportREMINDinputVarSet.R
\name{toolReportREMINDinputVarSet}
\alias{toolReportREMINDinputVarSet}
\title{Report detailed transport variable set}
\usage{
toolReportREMINDinputVarSet(
  fleetESdemand,
  fleetFEdemand,
  fleetEnergyIntensity,
  scenSpecLoadFactor,
  fleetCapCosts,
  combinedCAPEXandOPEX,
  scenSpecPrefTrends,
  scenSpecEnIntensity,
  initialIncoCosts,
  annualMileage,
  timeValueCosts,
  hybridElecShare,
  demScen,
  SSPscen,
  transportPolScen,
  timeResReporting,
  helpers
)

toolReportREMINDinputVarSet(
  fleetESdemand,
  fleetFEdemand,
  fleetEnergyIntensity,
  scenSpecLoadFactor,
  fleetCapCosts,
  combinedCAPEXandOPEX,
  scenSpecPrefTrends,
  scenSpecEnIntensity,
  initialIncoCosts,
  annualMileage,
  timeValueCosts,
  hybridElecShare,
  demScen,
  SSPscen,
  transportPolScen,
  timeResReporting,
  helpers
)
}
\arguments{
\item{fleetESdemand}{energy service demand on fleet level}

\item{fleetFEdemand}{final energy demand on fleet level}

\item{fleetEnergyIntensity}{energy intensity on fleet level}

\item{scenSpecLoadFactor}{scenario specific load factor data}

\item{fleetCapCosts}{annualized capital costs on fleet level}

\item{combinedCAPEXandOPEX}{CAPEX and OPEX on sales level in high temporal resolution}

\item{scenSpecPrefTrends}{scenario specific preference trends in high temporal resolution}

\item{scenSpecEnIntensity}{scenario specific energy intensity in high temporal resolution}

\item{initialIncoCosts}{initial inconvenience cost}

\item{annualMileage}{annual mileage in high temporal resolution}

\item{timeValueCosts}{time value cost equivalent in high temporal resolution}

\item{hybridElecShare}{share of electric driving for hybrid electric vehicles}

\item{demScen}{demand scenario}

\item{SSPscen}{SSP scenario}

\item{transportPolScen}{transport policy scenario}

\item{timeResReporting}{time resolution reporting}

\item{helpers}{list with helpers}
}
\value{
Detailed transport output variable set

REMIND/iterative EDGE-T input data
}
\description{
Report detailed transport variable set

Report REMIND/iterative EDGE-T input data
}
\author{
Johanna Hoppe
}
