% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReportsharesLDVtransport.R
\name{toolReportsharesLDVtransport}
\alias{toolReportsharesLDVtransport}
\title{Report FE shares of LDV in transport liquids}
\usage{
toolReportsharesLDVtransport(fleetFEdemand, timeResReporting, helpers)
}
\arguments{
\item{fleetFEdemand}{final energy demand on fleet level}

\item{timeResReporting}{time resolution reporting}

\item{helpers}{list of helpers}
}
\value{
Final energy shares of LDV in transport liquids [-]
}
\description{
Report FE shares of LDV in transport liquids
}
\author{
Johanna Hoppe
}
