% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateVariables.R
\name{aggregateVariables}
\alias{aggregateVariables}
\title{Aggregate variables}
\usage{
aggregateVariables(vars, mapAggregation, weight = NULL)
}
\arguments{
\item{vars}{Data.table with variables to aggregate in modified quitte style format}

\item{mapAggregation}{Map containing levels for aggregation in addition to the levels of the decision tree}

\item{weight}{Weight to aggregate variables}
}
\value{
Data.table with aggregated variables
}
\description{
This function aggregates a large set of variables according to the edge transport decision tree and
additional aggregation levels that are set manually.
The aggregation is applied to all variables simultaneously. Additional transport specific columns are
transferred into the variable column entry
}
\author{
Johanna Hoppe
}
