% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporttransport-package.R
\docType{package}
\name{reporttransport-package}
\alias{reporttransport}
\alias{reporttransport-package}
\title{reporttransport: Reporting package for edgeTransport}
\description{
This package contains edgeTransport-specific routines to report model results. The main functionality is to generate transport reporting variables in MIF format from a given edgeTransport model run folder or REMIND input data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/reporttransport}
}

}
\author{
\strong{Maintainer}: Johanna Hoppe \email{johanna.hoppe@pik-potsdam.de}

Authors:
\itemize{
  \item Jarusch Muessel \email{jarusch.muessel@pik-potsdam.de} (\href{https://orcid.org/0000-0002-1857-7866}{ORCID})
  \item Alex K. Hagen \email{alex.hagen@pik-potsdam.de} (\href{https://orcid.org/0000-0003-4793-8664}{ORCID})
}

}
\keyword{internal}
