% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReportsharesLDVtransport.R
\name{toolReportsharesLDVtransport}
\alias{toolReportsharesLDVtransport}
\title{Report FE shares of LDV in transport liquids}
\usage{
toolReportsharesLDVtransport(
  fleetFEdemand,
  timeResReporting,
  demScen,
  SSPscen,
  transportPolScen,
  helpers
)
}
\arguments{
\item{fleetFEdemand}{final energy demand on fleet level}

\item{timeResReporting}{time resolution reporting}

\item{demScen}{demand scenario}

\item{SSPscen}{SSP scenario}

\item{transportPolScen}{transport policy scenario}

\item{helpers}{list of helpers}
}
\value{
Final energy shares of LDV in transport liquids [-]
}
\description{
Report FE shares of LDV in transport liquids
}
\author{
Johanna Hoppe
}
