% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReportUE.R
\name{toolReportUE}
\alias{toolReportUE}
\title{Report variables in relation to the vehicle fleet.}
\usage{
toolReportUE(FEdemand, helpers)
}
\arguments{
\item{FEdemand}{xx}

\item{helpers}{xx}
}
\value{
xx
}
\description{
Report variables in relation to the vehicle fleet.
}
\author{
Johanna Hoppe
}
